[Refactor] SentRequestCard/ReceivedRequestCard 공통 베이스 컴포넌트 추출
- BaseRequestCard를 components/mypage에 생성하여 공통 로직 통합
- 유틸 함수(parseTime, formatShortDate, formatFullDate), 상수(TYPE_LABEL, STATUS_CONFIG) 통합
- 카드 레이아웃, 상세 섹션(요청유형, 기존/요청 근무시간, 휴게시간, 메모, 처리일시) 공통화
- 차이점은 props로 처리: headerSubtitle, extraDetailFields, actionButtons
- 거절 상태 색상을 근로자/고용주 동일하게 통일 (red + white)
